+++ src/components/place-card/place-card.tsx
@@ -1,20 +1,4 @@
-export const Place = {
-  Private: 'Private room',
-  Apartment: 'Apartment'
-} as const;
-
-type PlaceType = (typeof Place)[keyof typeof Place];
-
-export type Card = {
-  id: number;
-  isPremium: boolean;
-  imgSrc: string;
-  price: number;
-  inBookmarks: boolean;
-  rating: number;
-  name: string;
-  type: PlaceType;
-};
+import { Card } from "../../models";
 
 type PlaceCardProps = {
   card: Card;
